fix: prevent Mongoose models from leaking into client bundle (#336)
Root cause: TanStack Start keeps .validator() chains in the client bundle.
SUPPORTED_COLLECTIONS was derived from COLLECTION_REGISTRY (which references
Note.find), pulling Mongoose into the browser → "Class extends value undefined"
crash on the play route.

Two fixes:
1. Make SUPPORTED_COLLECTIONS a static literal instead of deriving it from
   COLLECTION_REGISTRY, breaking the transitive dependency on Mongoose models.
2. Move removeDocumentRefsFromScreens to gmscreens-helpers.ts so that notes.ts
   importing it doesn't force the bundler to pull in the full gmscreens module.
